Cоздание модуля вынесено в Assembly
