+++ homework-5/homework-5/Modules/BodyPicker/BodyPickerRouter.swift
@@ -8,25 +8,6 @@
 import UIKit
 
 final class BodyPickerRouter: BodyPickerRouterProtocol {
-    static func createBodyPickerViewController(with body: Body?) -> UIViewController {
-        let bodyPickerViewController = BodyPickerViewController()
-        
-        let interactor = BodyPickerInteractor()
-        let presenter = BodyPickerPresenter()
-        let router = BodyPickerRouter()
-        
-        bodyPickerViewController.presenter = presenter
-        bodyPickerViewController.selectedBody = body
-        
-        interactor.presenter = presenter
-        
-        presenter.viewController = bodyPickerViewController
-        presenter.interactor = interactor
-        presenter.router = router
-         
-        return bodyPickerViewController
-    }
-    
     func closeBodyPickerViewController(_ viewController: BodyPickerViewControllerProtocol) {
         guard let viewController = viewController as? UIViewController else { return }
