+++ homework-5/homework-5/Modules/CarDetail/CarDetailRouter.swift
@@ -8,27 +8,8 @@
 import UIKit
 
 final class CarDetailRouter: CarDetailRouterProtocol {
-    static func createCarDetailViewController(with car: Car?) -> UIViewController {
-        let carDetailViewController = CarDetailViewController()
-        
-        let interactor = CarDetailInteractor()
-        let presenter = CarDetailPresenter()
-        let router = CarDetailRouter()
-        
-        carDetailViewController.presenter = presenter
-        carDetailViewController.carToEdit = car
-        
-        interactor.presenter = presenter
-        
-        presenter.viewController = carDetailViewController
-        presenter.interactor = interactor
-        presenter.router = router
-         
-        return carDetailViewController
-    }
-    
     func openBodyPickerViewController(from viewController: CarDetailViewControllerProtocol, with body: Body?) {
-        let bodyPickerViewController = BodyPickerRouter.createBodyPickerViewController(with: body)
+        let bodyPickerViewController = Assembly.createBodyPickerViewController(with: body)
         
         if let bodyPickerViewController = bodyPickerViewController as? BodyPickerViewControllerProtocol {
             bodyPickerViewController.didSelectBody = { body in
