+++ homework-5/homework-5/Protocols/BodyPicker/BodyPickerRouterProtocol.swift
@@ -8,7 +8,5 @@
 import UIKit
 
 protocol BodyPickerRouterProtocol: AnyObject {
-    static func createBodyPickerViewController(with body: Body?) -> UIViewController
-    
     func closeBodyPickerViewController(_ viewController: BodyPickerViewControllerProtocol)
 }
