+++ homework-5/homework-5/SceneDelegate.swift
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         
         window = UIWindow(windowScene: windowScene)
-        window?.rootViewController = CarsRouter.createCarsNavigationController()
+        window?.rootViewController = Assembly.createCarsNavigationController()
         window?.makeKeyAndVisible()
     }
 }
